RestVerticle checks for implementation at startup RMB-805
Big refactor; all user requests handled in static class RestRouting.
Use Vert.x route for each of the routes offered, instead of
one big fat route function that does everything (slow and big).
This has been tested with mod-inventory-storage and all tests pass.
RMB returns 404 on path/method not found (actually this is thrown by Vert.x
itself now that only supported routes are registered).
